requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
